// packages/ui/src/components/Table2/hooks/types.ts
/*
 * Copyright 2025 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Return type for useTableAsyncData hook.
 * Provides all pagination information needed to feed TablePagination component.
 *
 * @public
 */
export interface UseTableAsyncDataResult {
  data: Record<string, any>[];
  loading: boolean;
  error: Error | null;
  fromCount: number;
  toCount: number;
  totalCount?: number;
  pageSize: number;
  onNextPage: () => void;
  onPreviousPage: () => void;
  onPageSizeChange: (pageSize: number) => void;
  isNextDisabled: boolean;
  isPrevDisabled: boolean;
  showPageSizeOptions: boolean;
}

/**
 * Return type for useTableData hook.
 * Provides all pagination information needed to feed TablePagination component.
 *
 * @public
 */
export interface UseTableDataResult {
  data: Record<string, any>[];
  loading: boolean;
  error: Error | null;
  totalCount: number;
  fromCount: number;
  toCount: number;
  pageSize: number;
  onNextPage: () => void;
  onPreviousPage: () => void;
  onPageSizeChange: (pageSize: number) => void;
  isNextDisabled: boolean;
  isPrevDisabled: boolean;
  showPageSizeOptions: boolean;
}